+++ fabric-item-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/item/v1/CustomDamageHandler.java
@@ -1,0 +1,11 @@
+package net.fabricmc.fabric.api.item.v1;
+
+import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+
+@FunctionalInterface
+public interface CustomDamageHandler {
+    int damage(ItemStack stack, int amount, LivingEntity entity,
+            EquipmentSlot slot, Runnable breakCallback);
+}