+++ fabric-item-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/item/v1/EnchantingContext.java
@@ -1,0 +1,6 @@
+package net.fabricmc.fabric.api.item.v1;
+
+public enum EnchantingContext {
+    ACCEPTABLE,
+    PRIMARY
+}