+++ fabric-item-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/item/v1/EnchantmentSource.java
@@ -1,0 +1,9 @@
+package net.fabricmc.fabric.api.item.v1;
+
+public enum EnchantmentSource {
+    VANILLA(true), MOD(true), DATA_PACK(false);
+
+    private final boolean builtin;
+    EnchantmentSource(boolean builtin) { this.builtin = builtin; }
+    public boolean isBuiltin() { return builtin; }
+}