+++ fabric-item-api-v1-bridge/src/main/java/net/fabricmc/fabric/api/item/v1/FabricTooltipType.java
@@ -1,0 +1,5 @@
+package net.fabricmc.fabric.api.item.v1;
+
+public interface FabricTooltipType {
+    default boolean shouldDisplayAllInformation() { return false; }
+}